refactor: 백엔드 코드 리펙토링
