+++ packages/backend/src/submodules/syncbot/submodules/log/log.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
-import { SyncbotService } from './syncbot.service';
-import { SyncbotController } from './syncbot.controller';
-import { SyncbotStreamModule } from './submodules/stream/syncbotStream.module';
+import { SyncBotLogController } from './log.controller';
+import { SyncbotLogService } from './log.service';
+import { UserService } from 'src/submodules/user/user.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import {
     CalendarEntity,
@@ -10,20 +10,18 @@ import {
     UserEntity,
 } from '@opize/calendar2notion-model';
 import { HttpModule } from '@nestjs/axios';
-import { UserService } from 'src/submodules/user/user.service';
 
 @Module({
-    controllers: [SyncbotController],
-    providers: [SyncbotService, UserService],
+    controllers: [SyncBotLogController],
+    providers: [SyncbotLogService, UserService],
     imports: [
-        SyncbotStreamModule,
         TypeOrmModule.forFeature([
-            SyncBotEntity,
             UserEntity,
+            SyncBotEntity,
             CalendarEntity,
             EventEntity,
         ]),
         HttpModule,
     ],
 })
-export class SyncbotModule {}
+export class SyncbotLogModule {}
