+++ packages/backend/src/submodules/syncbot/submodules/stream/syncbotStream.controller.ts
@@ -3,9 +3,6 @@ import {
     Get,
     Post,
     Body,
-    Patch,
-    Param,
-    Delete,
     Sse,
     Res,
     MessageEvent,
