+++ packages/backend/src/submodules/syncbot/types/storageMap.ts
@@ -1,0 +1,21 @@
+export type ManagerStorageMap = {
+    readonly prefix: string;
+    readonly startedAt: Date;
+    timeout: number;
+    stop: boolean;
+    readonly verizon: string;
+
+    readonly workerAmount: {
+        pro: number;
+        free: number;
+        sponsor: number;
+    };
+
+    work: {
+        [id: string]: {
+            loopId: string;
+            nowWorkUserId: number | undefined;
+            completedSyncCount: number;
+        };
+    };
+};
